requirements.txt
opencv-python-headless
pandas
numpy
pydicom
pillow
python-dotenv